+++ R/utilities-output-selections.R
@@ -56,3 +56,23 @@ clearOutputs <- function(simulation) {
   simulation$outputSelections$clear()
   invisible(simulation)
 }
+
+#' @param quantitiesOrPaths Quantity instances (element or vector) (typically
+#'   retrieved using `getAllQuantitiesMatching`) or quantity path (element or
+#'   vector) to add.
+#'
+#' @param simulation Instance of a simulation for which output selection should
+#'   be updated.
+#'
+#' @title Set outputs
+#' @description
+#' Sets the quantities as output into the  `simulation`.
+#'  The quantities can either be specified using explicit instances or using paths.
+#'  This function clears the output selection before adding the new quantities.
+#'  See `addOutputs` for adding quantities without clearing the output selection.
+#'  See `clearOutputs` for clearing the output selection without adding new quantities.
+#' @export
+setOutputs <- function(quantitiesOrPaths, simulation) {
+  clearOutputs(simulation)
+  addOutputs(quantitiesOrPaths, simulation)
+}